moved class RachioApiException to own class file (again)
