+++ src/main/java/org/openhab/binding/rachio/handler/RachioDeviceHandler.java
@@ -31,7 +31,7 @@ import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
 import org.openhab.binding.rachio.RachioBindingConstants;
 import org.openhab.binding.rachio.internal.RachioConfiguration;
-import org.openhab.binding.rachio.internal.api.RachioApi.RachioApiException;
+import org.openhab.binding.rachio.internal.api.RachioApiException;
 import org.openhab.binding.rachio.internal.api.RachioDevice;
 import org.openhab.binding.rachio.internal.api.RachioEvent;
 import org.openhab.binding.rachio.internal.api.RachioZone;
